+++ R/mod_graph.R
@@ -43,6 +43,7 @@ mod_graph_ui <- function(id, plotHeight, plotWidth) {
 #' @param ColorReference Color reference line color.
 #' @param ColorPoints Dots color.
 #' @param ColorMemorized Memorized subgroup dots color.
+#' @param ColorFactorial Factorial Context color,
 #' @param click_points_data Clicked point subgroup.
 #' @param select_points_data Selected point subgroup.
 #' @param factorial_context_data  Data frame with factorial context subgroups.
@@ -54,7 +55,6 @@ mod_graph_ui <- function(id, plotHeight, plotWidth) {
 #' @param xlabel Label for x-axis
 #' @param circlestyle Appearance of circles.
 #' @param grid Grid used in plot background.
-#' @param pickradius Radius for point click.
 #' @param memorized_Data Data frame with all memorized subgroups.
 #' @param memorized_labels_on_off If labels of memorized subgroups should be displayed.
 #' @param subTitle Subtitle for graph.
@@ -71,16 +71,20 @@ mod_graph_server <- function(
   XRange = input$XRange,
   plot_type = input$plot_type,
   point_size = input$pointsize,
+  LabelParent = FALSE,
+  LabelMemorized = FALSE,
+  LabelTabClicked = FALSE,
+  LabelFactCont = FALSE,
   color,
   ColorBGplot,
   ColorTabClicked,
   ColorReference,
   ColorPoints,
   ColorMemorized,
+  ColorParents,
+  ColorFactorial,
   click_points_data = NULL,
-  select_points_data = NULL,
-  factorial_context_data = NULL,
-  parents_data = NULL,
+  ColorCustomReference,
   x = input$x,
   y = input$y,
   key = input$key,
@@ -88,7 +92,6 @@ mod_graph_server <- function(
   xlabel = input$xlabel,
   circlestyle = input$circlestyle,
   grid = input$grid,
-  pickradius = input$pickradius,
   memorized_Data,
   memorized_labels_on_off = shiny::reactive({FALSE}),
   subTitle = NULL,
@@ -100,13 +103,16 @@ mod_graph_server <- function(
   show_favour_arrows,
   favour_direction,
   favour_verum_name = NULL,
-  favour_comparator_name = NULL
+  favour_comparator_name = NULL,
+  add_funnel = FALSE,
+  exclude_funnel = FALSE,
+  alpha_funnel = 0.1
   ) {
 
   xmin <- xmax <- point_color <- memorizedText <- SGID <- text <- font.col <- NULL
   ns <- session$ns
 
-  #### graph ####
+    #### graph ####
   output$graph <- shiny::renderPlot({
     #require an not empty y-range value
     if (!is.null(YRange())) {
@@ -129,26 +135,100 @@ mod_graph_server <- function(
       #add column with memorized subgroup label ("" for no label as default)
       data$memorizedText <- ""
 
-      if (dim(memorized_Data())[1] > 0) {
-        tmp <- memorized_Data()
-        tmp$memorizedText <- ""
+      ## Label factorial Context
+      if (LabelFactCont()) {
+        if (dim(data[data$point_color %in% c(
+          ColorFactorial(),
+          different_hues(ColorFactorial(), value = 50),
+          different_hues(ColorFactorial(), value = 100),
+          different_hues(ColorFactorial(), value = 150)
+        ),])[1] > 0) {
 
-        for (i in 1:dim(memorized_Data())[1]) {
+        tmp <- data[data$point_color %in% c(
+          ColorFactorial(),
+          different_hues(ColorFactorial(), value = 50),
+          different_hues(ColorFactorial(), value = 100),
+          different_hues(ColorFactorial(), value = 150)
+        ),]
+        tmp$memorizedText <- ""
+        for (i in 1:dim(tmp)[1]) {
           tmp[i,"memorizedText"] <- paste0(
-            "N: ", tmp[i,x()], "\n",
-            y(), ": ", tmp[i,y()],"\n",
             paste(
               names(
                 tmp[,results()$factors]
-              )[tmp[i,results()$factors]!="Not used"],
+              )[tmp[i,results()$factors] != "Not used"],
               ":",
-              tmp[i,results()$factors][tmp[i,results()$factors]!="Not used"],
+              tmp[i,results()$factors][tmp[i,results()$factors] != "Not used"],
               collapse = "\n")
           )
         }
         data <- merge(tmp[,-1], data, all = TRUE)
       }
+      }
 
+      ## Label Parents
+      if (LabelParent()) {
+        if (dim(data[data$point_color %in% c(ColorParents()),])[1] > 0) {
+
+          tmp <- data[data$point_color %in% c(ColorParents()),]
+          tmp$memorizedText <- ""
+          for (i in 1:dim(tmp)[1]) {
+            tmp[i,"memorizedText"] <- paste0(
+              paste(
+                names(
+                  tmp[,results()$factors]
+                )[tmp[i,results()$factors] != "Not used"],
+                ":",
+                tmp[i,results()$factors][tmp[i,results()$factors] != "Not used"],
+                collapse = "\n")
+            )
+          }
+          data <- merge(tmp[,-1], data, all = TRUE)
+        }
+      }
+      ## Label Clicked
+      if (LabelTabClicked()) {
+        if (dim(data[data$point_color == c(ColorTabClicked()),])[1] > 0) {
+          tmp <- data[data$point_color == c(ColorTabClicked()),]
+          tmp$memorizedText <- ""
+          for (i in 1:dim(data[data$point_color == c(ColorTabClicked()),])[1]) {
+            tmp[i,"memorizedText"] <- paste0(
+              paste(
+                names(
+                  tmp[,results()$factors]
+                )[tmp[i,results()$factors] != "Not used"],
+                ":",
+                tmp[i,results()$factors][tmp[i,results()$factors] != "Not used"],
+                collapse = "\n")
+            )
+          }
+          data <- merge(tmp[,-1], data, all = TRUE)
+        }
+      }
+
+      ## Label Memorized
+      if (LabelMemorized()) {
+        if (dim(memorized_Data())[1] > 0) {
+          tmp <- memorized_Data()
+          tmp$memorizedText <- ""
+          tmp$point_color <- ColorMemorized()
+
+          for (i in 1:dim(memorized_Data())[1]) {
+            tmp[i,"memorizedText"] <- paste0(
+              "N: ", tmp[i,x()], "\n",
+              y(), ": ", tmp[i,y()],"\n",
+              paste(
+                names(
+                  tmp[,results()$factors]
+                )[tmp[i,results()$factors]!="Not used"],
+                ":",
+                tmp[i,results()$factors][tmp[i,results()$factors]!="Not used"],
+                collapse = "\n")
+            )
+          }
+          data <- merge(tmp[,-1], data, all = TRUE)
+        }
+      }
       # check for completeness of the data
       if (y() %in% colnames(data)) {
         data <- data[complete.cases(data[[y()]]),]
@@ -192,6 +272,40 @@ mod_graph_server <- function(
           rect_data <- NULL
         }
 
+        if (add_funnel()) {
+          quantiles <- results()$funnel_quantiles
+          #lower line
+          approx_lower_fun <-approxfun(
+            quantiles[quantiles[,"alpha"] == ((as.numeric(alpha_funnel()))/2),][,"n"],
+            quantiles[quantiles[,"alpha"] == ((as.numeric(alpha_funnel()))/2),][,y()]
+          )
+          #upper
+          approx_higher_fun <-approxfun(
+            quantiles[quantiles[,"alpha"] == 1-((as.numeric(alpha_funnel()))/2),][,"n"],
+            quantiles[quantiles[,"alpha"] == 1-((as.numeric(alpha_funnel()))/2),][,y()]
+          )
+
+          data <- data %>%
+            dplyr::mutate(lower = approx_lower_fun(data[,x()]),upper = approx_higher_fun(data[,x()])) %>%
+            dplyr::mutate(outlier = (data[y()] <= lower | data[y()] > upper)) %>%
+            dplyr::mutate(outlier = ifelse(is.na(outlier), FALSE, outlier))
+
+
+          if(!is.null(XRange)) {
+            data <- data %>%
+            dplyr::mutate(
+              x_range_1 = XRange[1],
+              x_range_2 = XRange[2]
+            )
+          } else {
+           data <- data %>%
+            dplyr::mutate(
+              x_range_1 = min(data[[x()]], na.rm = TRUE),
+              x_range_2 = max(data[[x()]], na.rm = TRUE)
+            )
+          }
+        }
+
         size_tmp <- ifelse(circlestyle() == "standard", "point_size", "N.of.subjects")
 
         YRange <- YRange()
@@ -210,6 +324,7 @@ mod_graph_server <- function(
             !!rlang::sym(y()) <= YRange[2]
           )
 
+     data_tmp <- data
       p <- ggplot2::ggplot(
          data %>% dplyr::arrange(desc(point_color)),
          ggplot2::aes(
@@ -218,7 +333,6 @@ mod_graph_server <- function(
            label = memorizedText
           )
       )
-
       if (!is.null(XRange)) {
         p <- p + ggplot2::scale_x_continuous(
           limits = XRange,
@@ -247,6 +361,39 @@ mod_graph_server <- function(
           )
       }
 
+        if (add_funnel()) {
+       p <- p +
+        ggplot2::geom_ribbon(
+          data = data,
+          ggplot2::aes(ymax = upper, ymin = lower,xmax = x_range_2),
+          fill = ColorPoints(),
+          alpha = 0.2
+        )
+         if (exclude_funnel()) {
+
+          data <- data %>%
+            dplyr::filter(
+              outlier == TRUE |
+                ((outlier == FALSE & data$point_color != ColorPoints()) &
+                (outlier == FALSE & data$point_color != grDevices::adjustcolor(ColorPoints(), alpha = 0.75 )) &
+                (outlier == FALSE & data$point_color != grDevices::adjustcolor(ColorPoints(), alpha = 0.5 )) &
+                (outlier == FALSE & data$point_color != grDevices::adjustcolor(ColorPoints(), alpha = 0.25 )) &
+                (outlier == FALSE & data$point_color != grDevices::adjustcolor(ColorPoints(), alpha = 0.1 )))
+            )
+        } else {
+
+        data <- data %>%
+          dplyr::mutate(
+            point_color =
+              dplyr::case_when(
+                outlier == TRUE ~ point_color,
+                outlier == FALSE & data$point_color == ColorPoints() ~ point_color,
+                outlier == FALSE & data$point_color != ColorPoints()  ~ point_color,
+                is.na(outlier) ~ point_color
+              )
+          )
+        }
+      }
       if (is.null(subTitle) && xlabel()){
         p <- p +
           ggplot2::labs(
@@ -292,15 +439,33 @@ mod_graph_server <- function(
           )
       }
 
-    if (memorized_labels_on_off()) {
-      p <- p +
-        ggrepel::geom_text_repel(
-          colour = ColorMemorized(), max.overlaps = Inf,
-          min.segment.length = 0,
-          size = 5,
-          bg.r = 0.15
+     if (!exclude_funnel()) {
+
+        p <- p +
+          ggrepel::geom_label_repel(
+            colour = "white",
+            fill = data %>%
+              dplyr::arrange(desc(point_color)) %>%
+              dplyr::pull(point_color),
+            max.overlaps = Inf,
+            min.segment.length = 0,
+            size = 5#,
+            #bg.r = 0.15
           )
-      }
+     } else {
+       p <- p +
+       ggrepel::geom_label_repel(
+            colour = "white",
+            #fill = "#424242",
+            fill = data_tmp %>%
+              dplyr::arrange(desc(point_color)) %>%
+              dplyr::pull(point_color),
+            max.overlaps = Inf,
+            min.segment.length = 0,
+            size = 5
+          )
+     }
+
       p <- p + ggplot2::scale_colour_identity() +
         ggplot2::theme(
           panel.background = ggplot2::element_rect(color = NA, fill = NA),
@@ -342,7 +507,7 @@ mod_graph_server <- function(
               p <- p +
               ggplot2::geom_hline(
                 yintercept = value_custom_ref_line(),
-                color = "#00BCFFFF"
+                color = ColorCustomReference()
               )
               # reference value label in blue for reference line
                 p <- p +
@@ -351,7 +516,7 @@ mod_graph_server <- function(
                   y = value_custom_ref_line(),
                   label = value_custom_ref_line(),
                   color = "black",
-                  fill = "#00BCFFFF"
+                  fill = ColorCustomReference()
                 )
             }
           }
@@ -367,7 +532,7 @@ mod_graph_server <- function(
             ggplot2::annotate(geom = "label",
               x = ifelse(!is.null(XRange()),XRange[2],max(rect_data)),
               y = YRange[2],
-              label = ifelse(favour_direction(),ifelse(is.null(favour_verum_name),"Favours Verum",paste0("Favours ", favour_verum_name)),ifelse(is.null(favour_comparator_name),"Favour Comparator",paste0("Favour ",favour_comparator_name))),
+              label = ifelse(favour_direction(),ifelse(is.null(favour_verum_name),"Favour Verum",paste0("Favour ", favour_verum_name)),ifelse(is.null(favour_comparator_name),"Favour Comparator",paste0("Favour ",favour_comparator_name))),
               color = ifelse(favour_direction(),"black","white"),
               fill = ifelse(favour_direction(),"#89D329FF","#FF3162FF"),
               size = ggplot2::unit(6, "pt"),
@@ -376,7 +541,7 @@ mod_graph_server <- function(
             ggplot2::annotate(geom = "label",
               x = ifelse(!is.null(XRange()),XRange[2],max(rect_data)),#-ten_percent_x_range,
               y = YRange[1],
-              label = ifelse(favour_direction(),ifelse(is.null(favour_comparator_name),"Favours Comparator",paste0("Favours ",favour_comparator_name)), ifelse(is.null(favour_verum_name),"Favour Verum",paste0("Favour ", favour_verum_name))),
+              label = ifelse(favour_direction(),ifelse(is.null(favour_comparator_name),"Favour Comparator",paste0("Favour ",favour_comparator_name)), ifelse(is.null(favour_verum_name),"Favour Verum",paste0("Favour ", favour_verum_name))),
               color = ifelse(favour_direction(),"white", "black"),
               fill = ifelse(favour_direction(),"#FF3162FF", "#89D329FF"),
               size = ggplot2::unit(6, "pt"),
@@ -413,9 +578,28 @@ mod_graph_server <- function(
             }
           }
         }
+
+       if (add_funnel()) {
+          p <- p +
+            ggplot2::geom_point(
+              data = data %>%
+                dplyr::arrange(desc(point_color)) %>%
+                dplyr::filter(outlier == TRUE),
+              fill = data %>%
+                dplyr::arrange(desc(point_color)) %>%
+                dplyr::filter(outlier == TRUE) %>%
+                dplyr::pull(point_color),
+              shape = 21,
+              size = point_size(),
+              colour ="#ffffff30",
+              stroke = 0.9,
+              na.rm = TRUE,
+            )
+       }
         p
       }
     }
+
   })
 
   #click event handler
@@ -423,23 +607,20 @@ mod_graph_server <- function(
 
   shiny::observeEvent(c(input$plot_click), {
     curr_x <- shiny::req(x())
-    start_radius <- pickradius()
 
     clicked <- shiny::nearPoints(
       results()$sge[which(results()$sge$nfactors >= key()[1] & results()$sge$nfactors <= key()[2]),],
       input$plot_click,
       xvar = curr_x,
       yvar = y(),
-      threshold = start_radius,
       maxpoints = NULL
     )
-
     clicked <- subset(
       clicked,
       select = c("SGID", x = curr_x, y = y(), "nfactors", results()$factors)
     )
-
     click_points_data$xy <- clicked[, unlist(lapply(clicked, function(x) !all(is.na(x))))]
+    click_points_data$xy
   })
 
   shiny::observeEvent(c(input$plot_click), {
@@ -447,7 +628,6 @@ mod_graph_server <- function(
     all_points <- cbind(plot_point(), color = color(), stringsAsFactors = FALSE)
 
     colored_points <- all_points
-
     tmp_complement <- data.frame(
       plot_points_data_complement()["ID"],
       x = unname(plot_points_data_complement()["N.of.subjects.complement"]),
@@ -459,30 +639,39 @@ mod_graph_server <- function(
 
     click <- input$plot_click
     click$mapping <- list(xintercept = "xintercept", x = "x", y = "y")
-
     point <- nearPoints(colored_points, click)
-
   if (nrow(point) == 0) {
     selected_SGIDs$val <- NULL
     output$click_info <- shiny::renderUI({
       NULL
     })
   } else {
-    left_pct <- (click$coords_img$x - click$range$left) / (click$range$right - click$range$left)
-    top_pct <- (click$domain$top - ifelse(plot_type() == "lin", click$y, log10(click$y))) / (click$domain$top - click$domain$bottom)
+    # left_pct <- (click$coords_img$x - click$range$left) / (click$range$right - click$range$left)
+    # top_pct <- (click$domain$top - ifelse(plot_type() == "lin", click$y, log10(click$y))) / (click$domain$top - click$domain$bottom)
+    # left_px <- (click$range$left + left_pct * (click$range$right - click$range$left) / click$img_css_ratio$x) + 3
+    # top_px <- (click$range$top + top_pct * (click$range$bottom - click$range$top) / click$img_css_ratio$y) - 47
 
-    left_px <- (click$range$left + left_pct * (click$range$right - click$range$left) / click$img_css_ratio$x) + 3
-    top_px <- (click$range$top + top_pct * (click$range$bottom - click$range$top) / click$img_css_ratio$y) - 47
+    left_px <- click$coords_css$x
+    top_px  <- click$coords_css$y
 
     style <- paste0(
-      "position:absolute;
-       z-index:110;background-color: rgba(",
+      "position:absolute; z-index:110; ",
+      "background-color: rgba(",
         grDevices::col2rgb(ColorTabClicked())[1],",",
         grDevices::col2rgb(ColorTabClicked())[2],",",
         grDevices::col2rgb(ColorTabClicked())[3],",0.95); ",
-        "left:", left_px, "px; top:", top_px, "px; border: 0px;"
+      "left:", left_px + 0,
+      "px; top:", top_px + 0, "px;"
     )
 
+    # style <- paste0(
+    #   "position:absolute;
+    #    z-index:110;background-color: rgba(",
+    #     grDevices::col2rgb(ColorTabClicked())[1],",",
+    #     grDevices::col2rgb(ColorTabClicked())[2],",",
+    #     grDevices::col2rgb(ColorTabClicked())[3],",0.95); ",
+    #     "left:", left_px, "px; top:", top_px, "px; border: 0px;"
+    # )
     tmp <- results()$sge[results()$sge$SGID %in% point$ID, ]
 
     tmp$text <- apply(
@@ -490,7 +679,6 @@ mod_graph_server <- function(
       1,
       function(x){paste(paste0(names(which(x != "Not used")),":", x[which(x != "Not used")]), collapse = ", ")}
     )
-
     tmp <- tmp %>%
       dplyr::mutate(text2 = paste0("SGID:", SGID,", ", text))
 
@@ -528,7 +716,6 @@ mod_graph_server <- function(
   ## hover information window
   output$hover_info <- shiny::renderUI({
     shiny::req(input$plot_hover, plot_point())
-
     input$plot_hover
     all_points <- cbind(plot_point(), color = color(), stringsAsFactors = FALSE)
 
@@ -549,17 +736,16 @@ mod_graph_server <- function(
     point <- nearPoints(colored_points, hover)
 
     if (nrow(point) == 0) return(NULL)
-
-      left_pct <- (hover$coords_img$x - hover$range$left) / (hover$range$right - hover$range$left)
-
-      top_pct <- (hover$domain$top - ifelse(plot_type() == "lin", hover$y, log10(hover$y))) / (hover$domain$top - hover$domain$bottom)
-
-      left_px <- (hover$range$left + left_pct * (hover$range$right - hover$range$left) / hover$img_css_ratio$x) + 3
-      top_px <- (hover$range$top + top_pct * (hover$range$bottom - hover$range$top) / hover$img_css_ratio$y) + 3
+      # left_pct <- (hover$coords_img$x - hover$range$left) / (hover$range$right - hover$range$left)
+      # top_pct <- (hover$domain$top - ifelse(plot_type() == "lin", hover$y, log10(hover$y))) / (hover$domain$top - hover$domain$bottom)
+      # left_px <- (hover$range$left + left_pct * (hover$range$right - hover$range$left) / hover$img_css_ratio$x) + 3
+      # top_px <- (hover$range$top + top_pct * (hover$range$bottom - hover$range$top) / hover$img_css_ratio$y) + 3
+      left_px <- hover$coords_css$x
+      top_px <- hover$coords_css$y
 
       style <- paste0(
         "position:absolute;
-        z-index:100;background-color: rgba(",
+        z-index:100; pointer-events:none; background-color: rgba(",
           grDevices::col2rgb(ColorBGplot())[1],",",
           grDevices::col2rgb(ColorBGplot())[2],",",
           grDevices::col2rgb(ColorBGplot())[3],",0.95); ",
@@ -570,11 +756,15 @@ mod_graph_server <- function(
     res_and_color <- cbind(results()$sge[results()$sge$nfactors >= key()[1] & results()$sge$nfactors <= key()[2],], font.col = color())
     tmp <- res_and_color[res_and_color$SGID %in% point$ID, ]
 
-    tmp$text <- apply(
-      tmp[, results()$factors],
-      1,
-      function(x){paste(paste0(names(which(x != "Not used")),":", x[which(x != "Not used")]), collapse = ", ")}
-    )
+    if (length(results()$factors) == 1){
+      tmp$text <- paste0(results()$factors,":",tmp[,results()$factors])
+    } else {
+      tmp$text <- apply(
+        tmp[, results()$factors],
+        1,
+        function(x){paste(paste0(names(which(x != "Not used")),":", x[which(x != "Not used")]), collapse = ", ")}
+      )
+    }
     tmp2 <- tmp %>%
       dplyr::mutate(
         text2 = paste0("ID:", SGID,", ", text),
@@ -612,6 +802,7 @@ mod_graph_server <- function(
       )
     )
     }
+
   })
 
  return(